refactor(backend): extract Stripe event handlers + webhook router
The 610-line subscription.routes.ts mixed three concerns: HTTP handlers
for /me, /checkout, /portal; the Stripe webhook router with its raw
body, signature verification, and dedup-and-claim; and per-event
business logic flipping subscription state and emitting audit rows.
The reviewer flagged this as a layering issue and a testability
blocker.

Split into:
- domain/stripe-event-handlers.ts (509 lines) — payload helpers,
  applySubscriptionState, findLocalSubscriptionForUpdate, the per-event
  handlers, and a single handleStripeEvent(trx, stripe, event)
  dispatcher. Unknown event types fall through to a debug log so adding
  a new subscription in the dashboard doesn't break us. Also exposes
  claimStripeEvent for the atomic dedup logic.
- presentation/routes/subscription-webhook.routes.ts (121 lines) —
  HTTP-only: signature verify, claim, transaction wrapper around the
  domain dispatcher, and transient-vs-permanent error classification.
- presentation/routes/subscription.routes.ts (187 lines) — just /me,
  /checkout, /portal. Re-exports subscriptionWebhookRouter so the
  index.ts mount point doesn't need to change.

Tests pass unchanged after pointing the applySubscriptionState import
at the new domain module.
